+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/craft-ai/alpes/compare/v0.0.1...HEAD) ##
+## [Unreleased](https://github.com/craft-ai/alpes/compare/v0.0.2...HEAD) ##
+
+## [0.0.2](https://github.com/craft-ai/alpes/compare/v0.0.1...v0.0.2) - 2018-03-23 ##
 
 ### Added ###